исправил

+++ website/app/views/game_upload_form_views.py
@@ -6,19 +6,10 @@ from django.views import View
 from ..compiler import Compiler
 from ..game_upload_form import GameForm
 from ..models import Game
-
-LANGUAGES = {
-    'c++': 'cpp',
-    'c#': 'cs',
-    'c': 'c',
-    'python': 'py',
-    'javascript': 'js',
-    'java': 'Java',
-}
+from website.settings import SUPPORTED_LANGUAGES as LANGUAGES
 
 
 class GameUploadFormView(View):
-
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
@@ -30,12 +21,14 @@ class GameUploadFormView(View):
         })
 
     def post(self, request, *args, **kwargs):
-        game_form = GameForm(data=request.POST)
-        if game_form.is_valid():
-            try:
-                Game.objects.get(id=request.session['game_id']).delete()
-            except:
-                pass
+            game_form = GameForm(data=request.POST)
+
+            if game_form.is_valid():
+                try:
+                    if not request.session.get('game_been_uploaded'):
+                        Game.objects.get(id=request.session['game_id']).delete()
+                except:
+                    pass
 
             game_model = Game.objects.create(**game_form.cleaned_data)
             request.session['game_id'] = game_model.id
@@ -45,32 +38,31 @@ class GameUploadFormView(View):
             game_model.rules = request.FILES['rules']
             game_model.save()
 
-        ideal_solution = Compiler(
-            game_model.ideal_solution.path,
-            LANGUAGES[request.POST['ideal_solution_language']],
-            partial(self.notify, label='ideal_solution')
-        )
+            ideal_solution = Compiler(
+                game_model.ideal_solution.path,
+                request.POST['ideal_solution_language'],
+                callback=partial(self.notify, label='ideal_solution')
+            )
 
-        play = Compiler(
-            game_model.play.path,
-            LANGUAGES[request.POST['play_language']],
-            partial(self.notify, label='play')
-        )
+            play = Compiler(
+                game_model.play.path,
+                request.POST['play_language'],
+                callback=partial(self.notify, label='play')
+            )
 
-        visualiser = Compiler(
-            game_model.visualiser.path,
-            LANGUAGES[request.POST['visualiser_language']],
-            partial(self.notify, label='visualiser')
-        )
-        request.session['ideal_solution_report_id'] = None
-        request.session['play_report_id'] = None
-        request.session['visualiser_report_id'] = None
+            visualiser = Compiler(
+                game_model.visualiser.path,
+                request.POST['visualiser_language'],
+                callback=partial(self.notify, label='visualiser')
+            )
+            request.session['ideal_solution_report_id'] = None
+            request.session['play_report_id'] = None
+            request.session['visualiser_report_id'] = None
 
-        ideal_solution.compile()
-        play.compile()
-        visualiser.compile()
-
-        return redirect('game_upload_compilation')
+            ideal_solution.compile()
+            play.compile()
+            visualiser.compile()
+            return redirect('game_upload_compilation')
 
     def notify(self, report, label):
         if label == 'ideal_solution':
